+++ mayan/apps/common/dependencies.py
@@ -7,6 +7,7 @@ from mayan.apps.dependencies.environments import (
     environment_build, environment_development, environment_documentation,
     environment_documentation_override
 )
+from mayan.settings.literals import PYTHON_WHEEL_VERSION
 
 from .literals import DEFAULT_TX_PATH
 
@@ -40,7 +41,7 @@ PythonDependency(
         ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
         (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
         SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    ''', module=__name__, name='django', version_string='==3.2.16'
+    ''', module=__name__, name='django', version_string='==3.2.19'
 )
 PythonDependency(
     copyright_text='''
@@ -213,7 +214,9 @@ PythonDependency(
 )
 PythonDependency(
     environments=(environment_build, environment_documentation),
-    module=__name__, name='wheel', version_string='==0.37.1'
+    module=__name__, name='wheel', version_string='=={}'.format(
+        PYTHON_WHEEL_VERSION
+    )
 )
 
 # Documentation