+++ mayan/apps/document_states/workflow_actions.py
@@ -27,7 +27,7 @@ class DocumentPropertiesEditAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _('The new label to be assigned to the document.'),
                         BASE_WORKFLOW_TEMPLATE_STATE_ACTION_HELP_TEXT
                     )
@@ -42,7 +42,7 @@ class DocumentPropertiesEditAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _(
                             'The new description to be assigned to the '
                             'document.'
@@ -138,7 +138,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _('The URL to access.'),
                         BASE_WORKFLOW_TEMPLATE_STATE_ACTION_HELP_TEXT
                     )
@@ -154,7 +154,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _('Time in seconds to wait for a response.'),
                         BASE_WORKFLOW_TEMPLATE_STATE_ACTION_HELP_TEXT
                     )
@@ -169,7 +169,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _('A JSON document to include in the request.'),
                         BASE_WORKFLOW_TEMPLATE_STATE_ACTION_HELP_TEXT
                     )
@@ -184,7 +184,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _(
                             'Username to use for making the request with '
                             'HTTP Basic Auth.'
@@ -202,7 +202,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _(
                             'Password to use for making the request with '
                             'HTTP Basic Auth.'
@@ -220,7 +220,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _(
                             'The HTTP method to use for the request. '
                             'Typical choices are OPTIONS, HEAD, POST, GET, '
@@ -238,7 +238,7 @@ class HTTPAction(WorkflowAction):
             'kwargs': {
                 'initial_help_text': _(
                     format_lazy(
-                        '{}. {}',
+                        '{} {}',
                         _(
                             'Headers to send with the HTTP request. Must '
                             'be in JSON format.'