+++ mayan/apps/documents/tests/mixins/document_mixins.py
@@ -6,7 +6,7 @@ from django.db.models import Q
 from mayan.apps.converter.classes import Layer
 
 from ...literals import PAGE_RANGE_ALL
-from ...models import Document, DocumentType
+from ...models import Document
 
 from ..literals import (
     DEFAULT_DOCUMENT_STUB_LABEL, TEST_DOCUMENT_DESCRIPTION,
@@ -15,17 +15,23 @@ from ..literals import (
     TEST_FILE_SMALL_PATH
 )
 
+from .document_type_mixins import DocumentTypeTestMixin
+
 
 class DocumentAPIViewTestMixin:
     def _request_test_document_change_type_api_view(self):
         return self.post(
             viewname='rest_api:document-change-type', kwargs={
                 'document_id': self._test_document.pk
-            }, data={'document_type_id': self._test_document_types[1].pk}
+            }, data={
+                'document_type_id': self._test_document_types[1].pk
+            }
         )
 
     def _request_test_document_create_api_view(self):
-        pk_list = list(Document.objects.values_list('pk', flat=True))
+        pk_list = list(
+            Document.objects.values_list('pk', flat=True)
+        )
 
         response = self.post(
             viewname='rest_api:document-list', data={
@@ -67,7 +73,9 @@ class DocumentAPIViewTestMixin:
         return self.get(viewname='rest_api:document-list')
 
     def _request_test_document_upload_api_view(self):
-        pk_list = list(Document.objects.values_list('pk', flat=True))
+        pk_list = list(
+            Document.objects.values_list('pk', flat=True)
+        )
 
         with open(file=TEST_FILE_SMALL_PATH, mode='rb') as file_object:
             response = self.post(
@@ -87,7 +95,7 @@ class DocumentAPIViewTestMixin:
         return response
 
 
-class DocumentTestMixin:
+class DocumentTestMixin(DocumentTypeTestMixin):
     _test_document_count = 1
     _test_document_file_filename = TEST_FILE_SMALL_FILENAME
     _test_document_file_path = None
@@ -108,13 +116,10 @@ class DocumentTestMixin:
         self._test_document_file_pages = []
         self._test_document_id_list = []
         self._test_document_id_list_string = []
-        self._test_document_types = []
         self._test_document_versions = []
         self._test_document_version_pages = []
 
         if self.auto_create_test_document_type:
-            self._create_test_document_type()
-
             if self._test_document_count > 1:
                 if self.auto_upload_test_document:
                     self._upload_test_documents()
@@ -126,11 +131,19 @@ class DocumentTestMixin:
                 elif self.auto_create_test_document_stub:
                     self._create_test_document_stub()
 
-    def tearDown(self):
-        if self.auto_delete_test_document_type:
-            for document_type in DocumentType.objects.all():
-                document_type.delete()
-        super().tearDown()
+    def _calculate_test_document_file_path(self):
+        if not self._test_document_file_path:
+            self._test_document_file_path = os.path.join(
+                settings.BASE_DIR, 'apps', 'documents', 'tests', 'contrib',
+                'sample_documents', self._test_document_file_filename
+            )
+
+    def _calculate_test_document_path(self):
+        if not self._test_document_path:
+            self._test_document_path = os.path.join(
+                settings.BASE_DIR, 'apps', 'documents', 'tests', 'contrib',
+                'sample_documents', self._test_document_filename
+            )
 
     def _create_test_document_stub(self, document_type=None, label=None):
         if label is None:
@@ -153,26 +166,13 @@ class DocumentTestMixin:
         for index in range(count or self._test_document_count):
             self._create_test_document_stub()
 
-    def _create_test_document_type(self, label=None):
-        label = label or '{}_{}'.format(
-            TEST_DOCUMENT_TYPE_LABEL, len(self._test_document_types)
-        )
-
-        self._test_document_type = DocumentType.objects.create(label=label)
-        self._test_document_types.append(self._test_document_type)
-
-    def _calculate_test_document_path(self):
-        if not self._test_document_path:
-            self._test_document_path = os.path.join(
-                settings.BASE_DIR, 'apps', 'documents', 'tests', 'contrib',
-                'sample_documents', self._test_document_filename
-            )
-
-    def _calculate_test_document_file_path(self):
-        if not self._test_document_file_path:
-            self._test_document_file_path = os.path.join(
-                settings.BASE_DIR, 'apps', 'documents', 'tests', 'contrib',
-                'sample_documents', self._test_document_file_filename
+    def _create_test_documents(self, count=None):
+        for index in range(count or self._test_document_count):
+            self._upload_test_document(
+                label='{}_{}'.format(
+                    TEST_DOCUMENT_LABEL,
+                    len(self._test_documents)
+                )
             )
 
     def _upload_test_document(
@@ -229,15 +229,6 @@ class DocumentTestMixin:
                 setattr(self._test_document_version, key, value)
 
             self._test_document_version.save()
-
-    def _create_test_documents(self, count=None):
-        for index in range(count or self._test_document_count):
-            self._upload_test_document(
-                label='{}_{}'.format(
-                    TEST_DOCUMENT_LABEL,
-                    len(self._test_documents)
-                )
-            )
 
 
 class DocumentViewTestMixin: